Nowcast clinical laboratory percent positive flu data

This function provides a naive nowcasting method for clinical laboratory percent positive flu data. The methodology simply averages the last 4 weeks of available data and uses this average as the value for the number of weeks specified to replace. This is useful given that there is reporting lag in the NREVSS clinical laboratory percent positive flu data.

Usage

clin_nowcast(clin, weeks_to_replace = 1)

Arguments

clin

Data prepared with get_cdc_clin()

weeks_to_replace

Number of retrospective weeks to replace with nowcast; default is 1

Value

A tibble formatted the same as that returned with get_cdc_clin() but where the n most recent weeks (n="weeks_to_replace") have been nowcasted.

Examples

if (FALSE) {

## get data for Texas
tx_clin <-
get_cdc_clin(region = "state") %>%
dplyr::filter(location == "48")

## look at most recent observations
tx_clin %>%
dplyr::arrange(week_start) %>%
tail()

## now augment with default 1 week nowcast
tx_clin %>%
clin_nowcast(., weeks_to_replace = 1) %>%
dplyr::arrange(week_start) %>%
tail()

## and again augmented with 2 week nowcast instead
tx_clin %>%
 clin_nowcast(., weeks_to_replace = 2) %>%
 dplyr::arrange(week_start) %>%
 tail()

}