Replace all ILInet data with nowcast data for a state

Replaces weighted_ili from get_cdc_ili with nowcast data from get_nowcast_ili for all dates for a specified location. This is useful for getting data for states where most or all ILI data is missing (e.g., Florida).

Usage

state_replace_ili_nowcast_all(ilidat, state, impute = TRUE, ...)

Arguments

ilidat

Data from get_cdc_ili.

state

Two-letter state abbreviation to replace completely

impute

Logical; try to mean impute missing values using the immediately preceding and following values. See examples.

...

Other arguments passed to get_nowcast_ili, e.g. boundatzero, which is TRUE by default.

Value

The same as the ilidat input, but with state's data from get_cdc_ili replaced by nowcast data from get_nowcast_ili.

Details

This only replaces instances of weighted_ili in the specified state where weighted_ili is NA. Most ILI data from FL is missing, but not all.

See also

Examples

if (FALSE) {
ilidat <- get_cdc_ili(years=2020)
ilidat <-
  ilidat %>%
  dplyr::filter(location=="US" | abbreviation=="VA" | abbreviation=="FL") %>%
  dplyr::group_by(location) %>%
  dplyr::slice_max(week_start, n=4) %>%
  dplyr::select(location:weighted_ili) %>%
  dplyr::arrange(location, epiyear, epiweek)
ilidat
state_replace_ili_nowcast_all(ilidat, state="FL")
# Example with Florida, which has a negative value for nowcasted ILI
ilidat <- get_cdc_ili(years=2019)
ilidat <- ilidat %>%
  dplyr::filter(location=="US" | abbreviation=="VA" | abbreviation=="FL") %>%
  dplyr::filter(epiyear==2020 & epiweek %in% c(20, 21, 22)) %>%
  dplyr::select(location:weighted_ili) %>%
  dplyr::arrange(location, epiyear, epiweek)
ilidat
# defaults to bound at zero
state_replace_ili_nowcast_all(ilidat, state="FL")
# show results when you don't bound at zero
state_replace_ili_nowcast_all(ilidat, state="FL", boundatzero=FALSE)
# example with missing data in florida
ilidat <- get_cdc_ili(region=c("national","state"), years=2019:lubridate::year(lubridate::today()))
ilidat <- ilidat %>%
  dplyr::filter(abbreviation=="FL") %>%
  dplyr::filter(week_start>="2020-12-13" & week_start<="2021-01-10")
ilidat
state_replace_ili_nowcast_all(ilidat, state="FL")
state_replace_ili_nowcast_all(ilidat, state="FL", impute=FALSE)
}