Forecast ILI

Forecasts ILI up to specified weeks in the future. Used in downstream modeling.

Usage

forecast_ili(
  ilidat,
  horizon = 4L,
  trim_date = NULL,
  models = list(arima = "PDQ(0,0,0)+pdq(1:2,0:2,0)"),
  ...
)

Arguments

ilidat

Data returned from get_cdc_ili.

horizon

Optional horizon periods through which the forecasts should be generated; default is 4

trim_date

Earliest start date you want to use for ILI data. Default NULL doesn't trim.

models

The list of model parameters passed to ts_fit_forecast. Defaults to list(arima="PDQ(0,0,0)+pdq(1:2,0:2,0)". See help for ts_fit_forecast.

...

further arguments passed to ts_fit_forecast (which are then passed to fable::ARIMA.

Value

A named list containing:

ilidat: The data sent into the function filtered to the location and the trim_date. Select columns returned.

ilidat_tsibble: The tsibble class object returned by running make_tsibble on the data above.

ili_fit: The fit from fabletools::model.

ili_forecast: The forecast from fabletools::forecast at the specified horizon.

ili_future: The horizon-number of weeks of ILI data forecasted into the future.

ili_bound: The data in 1 bound to the data in 5.

arima_params: A tibble with ARIMA model parameters for each location (if type="arima").

locstats: A tibble with missing data information on all locations.

removed: A tibble with locations removed because of high missing ILI data.

Details

Currently limited to one location only.

Examples

if (FALSE) {
# Get data
ilidat <- get_cdc_ili(region = c("national", "state", "hhs"),
                      years = 2010:lubridate::year(lubridate::today()))

# Using data only from march 2020 forward, for US only
ilidat_us <- ilidat %>% dplyr::filter(location=="US")
# Replace most recent week with nowcast data, and nowcast last week
ilidat_us <- ilidat_us %>% replace_ili_nowcast(weeks_to_replace=1)
ilifor_us <- forecast_ili(ilidat_us, horizon=4L, trim_date="2020-03-01")
ilifor_us$ili_fit
ilifor_us$arima_params
ilifor_us$ili_forecast
head(ilifor_us$ili_bound)
tail(ilifor_us$ili_bound, 10)
# Plot
library(dplyr)
library(ggplot2)
theme_set(theme_classic())
ilifor_us$ili_bound %>%
  mutate(date=mmwr_week_to_date(epiyear, epiweek)) %>%
  filter(date>"2021-03-01") %>%
  ggplot(aes(date, ili)) +
  geom_line(lwd=.3, alpha=.5) +
  geom_point(aes(col=forecasted), size=2)

# At the state level
ilidat_st <- ilidat %>% dplyr::filter(region_type=="States")
ilifor_st <- forecast_ili(ilidat_st, horizon=4L, trim_date="2019-01-01",
                          models=list(ets="season(method='N')"))
ilifor_st$ili_fit
ilifor_st$arima_params
ilifor_st$ili_forecast
head(ilifor_us$ili_bound)
tail(ilifor_us$ili_bound, 10)
# Plot
library(dplyr)
library(ggplot2)
theme_set(theme_classic())
ilifor_st$ili_bound %>%
  mutate(date=mmwr_week_to_date(epiyear, epiweek)) %>%
  filter(date>"2021-08-01") %>%
  ggplot(aes(date, ili, col=forecasted)) +
  geom_line(lwd=.3) +
  geom_point(aes(col=forecasted), size=.7) +
  facet_wrap(~abbreviation, scale="free_y")

## At the HHS regional level
ilidat_hhs <- ilidat %>% dplyr::filter(region_type=="HHS Regions")
ilifor_hhs <- forecast_ili(ilidat_hhs, horizon=4L, trim_date="2020-03-01")
ilifor_hhs$ili_fit
ilifor_hhs$arima_params
ilifor_hhs$ili_forecast
head(ilifor_us$ili_bound)
tail(ilifor_us$ili_bound, 10)
# Plot
library(dplyr)
library(ggplot2)
theme_set(theme_classic())
ilifor_hhs$ili_bound %>%
  mutate(date=mmwr_week_to_date(epiyear, epiweek)) %>%
  filter(date>"2021-08-01") %>%
  ggplot(aes(date, ili, col=forecasted)) +
  geom_line(lwd=.3) +
  geom_point(aes(col=forecasted), size=.7) +
  facet_wrap(~abbreviation, scale="free_y")
}